+++ NEWS.md
@@ -1,6 +1,18 @@
 # EMLassemblyline 2.99.0
 
+#### 2019-03-14
+* __Bug fix:__ Add `//otherEntity/physical/distribution/online/url`
+
+#### 2019-03-11
+* __Refactor:__ Rename arguments
+    * `data.files` -> `data.table`
+    * `data.files.description` -> `data.table.description`
+    * `data.files.quote.character` -> `data.table.quote.character`
+    * `data.files.url` -> `data.url`
+
 #### 2019-03-02
-* Reuse `<entityDescription>` for `<entityName>`
-* Add `other.entity` and `other.entity.description` arguments to `make_eml`
-* Add change log
+* __Bug fix:__ Use arguments `data.table.description` and `other.entity.description` for `//dataTable/entityName` and `//otherEntity/entityName`, respectively.
+* __Refactor:__ Rename arguments 
+    * `zip.dir` -> `other.entity`
+    * `zip.dir.description` -> `other.entity.description`
+* __Enhancement:__ Add change log